fix overflow warning

+++ src/core/image.h
@@ -13,11 +13,11 @@ struct Image {
 	uint8_t*		data{ NULL };
 	size_t			size{ 0 };
 	int				channels{};
-	int				w{};
-	int				h{};
+	uint64_t				w{};
+	uint64_t				h{};
 
 	Image(const char* filename);
-	Image(int w, int h, int channels);
+	Image(uint64_t w, uint64_t h, int channels);
 	Image(const Image& img);
 	~Image();
 
@@ -30,7 +30,7 @@ struct Image {
 	Image& colorMask(float r, float g, float b, float a);
 
 	Image& overlay(const Image& source, int x, int y);
-	Image& resize(uint16_t width, uint16_t height);
+	Image& resize(uint32_t width, uint32_t height);
 	Image& addShadow(float intensity, uint16_t radius, float opacity);
 
 	Image& crop(uint16_t cx, uint16_t cy, uint16_t cw, uint16_t ch);
